backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "md2deck"
version = "0.1.0"
description = "Markdown to PowerPoint pipeline with slide-master-aware storytelling."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "markdown-it-py>=3.0.0",
  "matplotlib>=3.8.0",
  "python-pptx>=0.6.23",
  "PyYAML>=6.0.1",
  "fastapi>=0.109.0",
  "uvicorn>=0.27.0",
  "python-multipart>=0.0.9",
  "aiofiles>=23.2.1",
  "httpx>=0.26.0",
  "python-dotenv>=1.0.0",
  "google-generativeai>=0.8.0",
]

[project.scripts]
md2deck = "md2deck.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
